refactor: Put GUI related classes into package dude.gui
