+++ build.gradle
@@ -44,7 +44,7 @@ test {
 }
 
 application {
-    mainClass.set("dude.Launcher")
+    mainClass.set("dude.gui.Launcher")
 }
 
 shadowJar {
