+++ src/main/java/dude/gui/MainWindow.java
@@ -1,5 +1,6 @@
-package dude;
+package dude.gui;
 
+import dude.Dude;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ScrollPane;
